refactor: 로또 랜덤 뽑기 기능을 Lotto 클래스에서 MyLotto 클래스로 이동
