+++ src/models/Lotto.js
@@ -1,5 +1,3 @@
-import { Random } from '@woowacourse/mission-utils';
-import NUMBER from '../constants/Number';
 import OutputView from '../views/OutputView';
 
 class Lotto {
@@ -7,24 +5,7 @@ class Lotto {
 
   constructor(numbers) {
     this.#numbers = numbers;
-    this.#checkNumbersEmpty();
     this.#sortNumbers();
-  }
-
-  #checkNumbersEmpty() {
-    if (!this.#numbers) {
-      this.#numbers = this.#pickRandomNumbers();
-    }
-  }
-
-  #pickRandomNumbers() {
-    const randomNumbers = Random.pickUniqueNumbersInRange(
-      NUMBER.MIN_IN_RANGE,
-      NUMBER.MAX_IN_RANGE,
-      NUMBER.LOTTO_NUMBER_OF_NUMBERS,
-    );
-
-    return randomNumbers;
   }
 
   #sortNumbers() {
