+++ src/models/MyLotto.js
@@ -1,4 +1,6 @@
+import { Random } from '@woowacourse/mission-utils';
 import Lotto from './Lotto';
+import NUMBER from '../constants/Number';
 
 class MyLotto {
   #purchaseAmount;
@@ -12,8 +14,19 @@ class MyLotto {
 
   #createLottoNumbersList(lottoCount) {
     Array.from({ length: lottoCount }).forEach(() => {
-      this.#lottoNumbersList.push(new Lotto());
+      const numbers = this.#pickRandomNumbers();
+      this.#lottoNumbersList.push(new Lotto(numbers));
     });
+  }
+
+  #pickRandomNumbers() {
+    const randomNumbers = Random.pickUniqueNumbersInRange(
+      NUMBER.MIN_IN_RANGE,
+      NUMBER.MAX_IN_RANGE,
+      NUMBER.LOTTO_NUMBER_OF_NUMBERS,
+    );
+
+    return randomNumbers;
   }
 
   showMyLottoNumbers() {
